Securing Systems Against Log4Shell Exploits
The Log4Shell vulnerability is a weakness in Log4j where if someone sends a specially crafted message, Log4j tries to do something dangerous, it goes out to another computer (over the internet) to load and run some code it finds there. This happens because Log4j “interprets” parts of the message as instructions to load code remotely, this is called a JNDI lookup.
If the app or server is not protected properly, this lets an attacker run their own code on your server without permission. 

I experimented with this by simulating a vulnerable environment, exploiting it, then securing it, and testing.
I used AWS EC2 instance, Win SCP, and powershell to run this experiment. 
I set up Docker instance and deployed Java app with a vulnerable Log4j version (used Docker Compose). Exploited the vulnerability, used curl command to send the malicious JNDI payload. It was received - acknowledged which confirmed the vulnerability. 
Key steps I took to mitigate vulnerability: 
1.	Updating Log4j to 2.17.0
The older version had the flaw and hackers could trick into running code form the internet. The new version has fixed this problem making it harder for hackers to take advantage of this vulnerability. 
2.	Added input validation
a.	Switched @GetMapping to @PostMapping
This change makes the app safer by requiring users to send date in a more secure way – in the body of the message versus address sbar. 
b.	Used @RequestBody instead of query parameters 
Now the app reads the entire message body, which gives better control and visibility over what people are sending. 
c.	Checked for hacking patters like ${jndi
This filters incoming messages that may include a suspicious pattern and blocks them. 

Together these steps will help mitigate hackers ability to infiltrate the web application. This repo has 2 versions of the file to demonstrate how the code was changed to mitigate the problem.  